added unsigned -> signed cast

+++ src/cch_order.h
@@ -294,7 +294,7 @@ namespace cch_order{
 			}
 		);
 		
-		if(separator.size() == node_count){
+		if((int)separator.size() == node_count){
 			keep_arc_flag.fill(false);
 		}
